+++ src/main/java/nextstep/app/FilterConfig.java
@@ -2,21 +2,21 @@ package nextstep.app;
 
 import nextstep.app.domain.Member;
 import nextstep.app.domain.MemberRepository;
-import nextstep.security.AuthenticationManager;
-import nextstep.security.DaoAuthenticationProvider;
-import nextstep.security.DefaultSecurityFilterChain;
-import nextstep.security.FilterChainProxy;
-import nextstep.security.ProviderManager;
-import nextstep.security.SecurityFilterChain;
-import nextstep.security.context.HttpSessionSecurityContextRepository;
-import nextstep.security.filter.BasicAuthenticationFilter;
-import nextstep.security.filter.FormLoginAuthenticationFilter;
-import nextstep.security.UserDetails;
-import nextstep.security.UserDetailsService;
-import nextstep.security.filter.SecurityContextHolderFilter;
-import nextstep.security.password.PasswordEncoder;
-import nextstep.security.password.RawPasswordEncoder;
-import nextstep.security.util.matcher.AnyRequestMatcher;
+import nextstep.authentication.AuthenticationManager;
+import nextstep.authentication.DaoAuthenticationProvider;
+import nextstep.authentication.DefaultSecurityFilterChain;
+import nextstep.authentication.FilterChainProxy;
+import nextstep.authentication.ProviderManager;
+import nextstep.authentication.SecurityFilterChain;
+import nextstep.authentication.context.HttpSessionSecurityContextRepository;
+import nextstep.authentication.filter.BasicAuthenticationFilter;
+import nextstep.authentication.filter.FormLoginAuthenticationFilter;
+import nextstep.authentication.UserDetails;
+import nextstep.authentication.UserDetailsService;
+import nextstep.authentication.filter.SecurityContextHolderFilter;
+import nextstep.authentication.password.PasswordEncoder;
+import nextstep.authentication.password.RawPasswordEncoder;
+import nextstep.authentication.util.matcher.AnyRequestMatcher;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.filter.DelegatingFilterProxy;
