+++ src/main/java/nextstep/authentication/Authentication.java
@@ -1,0 +1,10 @@
+package nextstep.authentication;
+
+public interface Authentication {
+
+    String getPrincipal();
+
+    String getCredentials();
+
+    boolean isAuthenticated();
+}
