+++ src/main/java/nextstep/authentication/AuthenticationManager.java
@@ -1,0 +1,8 @@
+package nextstep.authentication;
+
+import nextstep.authentication.exception.AuthenticationException;
+
+public interface AuthenticationManager {
+
+    Authentication authenticate(Authentication authentication) throws AuthenticationException;
+}
