+++ src/main/java/nextstep/authentication/SecurityFilterChain.java
@@ -1,0 +1,13 @@
+package nextstep.authentication;
+
+import jakarta.servlet.Filter;
+import jakarta.servlet.http.HttpServletRequest;
+
+import java.util.List;
+
+public interface SecurityFilterChain {
+
+    boolean matches(HttpServletRequest request);
+
+    List<Filter> getFilters();
+}
