+++ src/main/java/nextstep/authentication/UserDetails.java
@@ -1,0 +1,7 @@
+package nextstep.authentication;
+
+public interface UserDetails {
+    String getUsername();
+
+    String getPassword();
+}
