+++ src/main/java/nextstep/authentication/UserDetailsService.java
@@ -1,0 +1,5 @@
+package nextstep.authentication;
+
+public interface UserDetailsService {
+    UserDetails loadUserByUsername(String username);
+}
