+++ src/main/java/nextstep/authentication/context/SecurityContext.java
@@ -1,0 +1,10 @@
+package nextstep.authentication.context;
+
+import nextstep.authentication.Authentication;
+
+public interface SecurityContext {
+
+    Authentication getAuthentication();
+
+    void setAuthentication(Authentication authentication);
+}
