+++ src/main/java/nextstep/authentication/context/SecurityContextHolder.java
@@ -1,0 +1,27 @@
+package nextstep.authentication.context;
+
+public class SecurityContextHolder {
+
+    private static final ThreadLocal<SecurityContext> contextHolder = new ThreadLocal<>();
+
+    public static SecurityContext getContext() {
+        SecurityContext context = contextHolder.get();
+        if (context == null) {
+            context = createEmptyContext();
+            contextHolder.set(context);
+        }
+        return context;
+    }
+
+    public static void setContext(SecurityContext context) {
+        contextHolder.set(context);
+    }
+
+    public static void clearContext() {
+        contextHolder.remove();
+    }
+
+    public static SecurityContext createEmptyContext() {
+        return new SecurityContextImpl();
+    }
+}
