+++ src/main/java/nextstep/authentication/context/SecurityContextImpl.java
@@ -1,0 +1,25 @@
+package nextstep.authentication.context;
+
+import nextstep.authentication.Authentication;
+
+public class SecurityContextImpl implements SecurityContext {
+
+    private Authentication authentication;
+
+    public SecurityContextImpl() {
+    }
+
+    public SecurityContextImpl(Authentication authentication) {
+        this.authentication = authentication;
+    }
+
+    @Override
+    public Authentication getAuthentication() {
+        return authentication;
+    }
+
+    @Override
+    public void setAuthentication(Authentication authentication) {
+        this.authentication = authentication;
+    }
+}
