+++ src/main/java/nextstep/authentication/context/SecurityContextRepository.java
@@ -1,0 +1,11 @@
+package nextstep.authentication.context;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+public interface SecurityContextRepository {
+
+    SecurityContext loadContext(HttpServletRequest request);
+
+    void saveContext(SecurityContext context, HttpServletRequest request, HttpServletResponse response);
+}
