+++ src/main/java/nextstep/authentication/exception/ProviderNotFoundException.java
@@ -1,0 +1,8 @@
+package nextstep.authentication.exception;
+
+public class ProviderNotFoundException extends AuthenticationException {
+
+    public ProviderNotFoundException(String message) {
+        super(message);
+    }
+}
