+++ src/main/java/nextstep/authentication/password/PasswordEncoder.java
@@ -1,0 +1,8 @@
+package nextstep.authentication.password;
+
+public interface PasswordEncoder {
+
+    String encode(String rawPassword);
+
+    boolean matches(String rawPassword, String encodedPassword);
+}
