+++ src/main/java/nextstep/authentication/password/RawPasswordEncoder.java
@@ -1,0 +1,14 @@
+package nextstep.authentication.password;
+
+public class RawPasswordEncoder implements PasswordEncoder {
+
+    @Override
+    public String encode(String rawPassword) {
+        return rawPassword;
+    }
+
+    @Override
+    public boolean matches(String rawPassword, String encodedPassword) {
+        return rawPassword.equals(encodedPassword);
+    }
+}
