+++ src/main/java/nextstep/authentication/util/matcher/RequestMatcher.java
@@ -1,0 +1,8 @@
+package nextstep.authentication.util.matcher;
+
+import jakarta.servlet.http.HttpServletRequest;
+
+public interface RequestMatcher {
+
+    boolean matches(HttpServletRequest request);
+}
