--- src/main/java/nextstep/security/Authentication.java
@@ -1,10 +1,0 @@
-package nextstep.security;
-
-public interface Authentication {
-
-    String getPrincipal();
-
-    String getCredentials();
-
-    boolean isAuthenticated();
-}
