--- src/main/java/nextstep/security/AuthenticationManager.java
@@ -1,8 +1,0 @@
-package nextstep.security;
-
-import nextstep.security.exception.AuthenticationException;
-
-public interface AuthenticationManager {
-
-    Authentication authenticate(Authentication authentication) throws AuthenticationException;
-}
