--- src/main/java/nextstep/security/AuthenticationProvider.java
@@ -1,10 +1,0 @@
-package nextstep.security;
-
-import nextstep.security.exception.AuthenticationException;
-
-public interface AuthenticationProvider {
-
-    Authentication authenticate(Authentication authentication) throws AuthenticationException;
-
-    boolean supports(Class<?> authentication);
-}
