--- src/main/java/nextstep/security/DefaultSecurityFilterChain.java
@@ -1,28 +1,0 @@
-package nextstep.security;
-
-import jakarta.servlet.Filter;
-import jakarta.servlet.http.HttpServletRequest;
-import nextstep.security.util.matcher.RequestMatcher;
-
-import java.util.List;
-
-public class DefaultSecurityFilterChain implements SecurityFilterChain {
-
-    private final RequestMatcher requestMatcher;
-    private final List<Filter> filters;
-
-    public DefaultSecurityFilterChain(RequestMatcher requestMatcher, List<Filter> filters) {
-        this.requestMatcher = requestMatcher;
-        this.filters = filters;
-    }
-
-    @Override
-    public boolean matches(HttpServletRequest request) {
-        return this.requestMatcher.matches(request);
-    }
-
-    @Override
-    public List<Filter> getFilters() {
-        return this.filters;
-    }
-}
