--- src/main/java/nextstep/security/ProviderManager.java
@@ -1,46 +1,0 @@
-package nextstep.security;
-
-import nextstep.security.exception.AuthenticationException;
-import nextstep.security.exception.ProviderNotFoundException;
-
-import java.util.List;
-
-public class ProviderManager implements AuthenticationManager {
-
-    private final List<AuthenticationProvider> providers;
-
-    public ProviderManager(List<AuthenticationProvider> providers) {
-        this.providers = providers;
-    }
-
-    @Override
-    public Authentication authenticate(Authentication authentication) throws AuthenticationException {
-
-        Class<? extends Authentication> toTest = authentication.getClass();
-        Authentication result = null;
-        AuthenticationException lastException = null;
-
-        for (AuthenticationProvider provider : this.providers) {
-            if (provider.supports(toTest)) {
-
-                try {
-                    result = provider.authenticate(authentication);
-                    if (result != null) {
-                        break;
-                    }
-                } catch (AuthenticationException ex) {
-                    lastException = ex;
-                }
-            }
-        }
-
-        if (result != null) {
-            return result;
-        }
-        if (lastException == null) {
-            lastException = new ProviderNotFoundException(String.format("No AuthenticationProvider found for %s", toTest.getName()));
-        }
-
-        throw lastException;
-    }
-}
