--- src/main/java/nextstep/security/SecurityFilterChain.java
@@ -1,13 +1,0 @@
-package nextstep.security;
-
-import jakarta.servlet.Filter;
-import jakarta.servlet.http.HttpServletRequest;
-
-import java.util.List;
-
-public interface SecurityFilterChain {
-
-    boolean matches(HttpServletRequest request);
-
-    List<Filter> getFilters();
-}
