--- src/main/java/nextstep/security/UserDetails.java
@@ -1,7 +1,0 @@
-package nextstep.security;
-
-public interface UserDetails {
-    String getUsername();
-
-    String getPassword();
-}
