--- src/main/java/nextstep/security/UserDetailsService.java
@@ -1,5 +1,0 @@
-package nextstep.security;
-
-public interface UserDetailsService {
-    UserDetails loadUserByUsername(String username);
-}
