--- src/main/java/nextstep/security/context/SecurityContext.java
@@ -1,10 +1,0 @@
-package nextstep.security.context;
-
-import nextstep.security.Authentication;
-
-public interface SecurityContext {
-
-    Authentication getAuthentication();
-
-    void setAuthentication(Authentication authentication);
-}
