--- src/main/java/nextstep/security/context/SecurityContextImpl.java
@@ -1,25 +1,0 @@
-package nextstep.security.context;
-
-import nextstep.security.Authentication;
-
-public class SecurityContextImpl implements SecurityContext {
-
-    private Authentication authentication;
-
-    public SecurityContextImpl() {
-    }
-
-    public SecurityContextImpl(Authentication authentication) {
-        this.authentication = authentication;
-    }
-
-    @Override
-    public Authentication getAuthentication() {
-        return authentication;
-    }
-
-    @Override
-    public void setAuthentication(Authentication authentication) {
-        this.authentication = authentication;
-    }
-}
