--- src/main/java/nextstep/security/exception/AuthenticationException.java
@@ -1,19 +1,0 @@
-package nextstep.security.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(code = HttpStatus.UNAUTHORIZED)
-public class AuthenticationException extends RuntimeException {
-
-    public AuthenticationException() {
-    }
-
-    public AuthenticationException(String msg, Throwable cause) {
-        super(msg, cause);
-    }
-
-    public AuthenticationException(String msg) {
-        super(msg);
-    }
-}
