--- src/main/java/nextstep/security/exception/ProviderNotFoundException.java
@@ -1,8 +1,0 @@
-package nextstep.security.exception;
-
-public class ProviderNotFoundException extends AuthenticationException {
-
-    public ProviderNotFoundException(String message) {
-        super(message);
-    }
-}
