--- src/main/java/nextstep/security/filter/FormLoginAuthenticationFilter.java
@@ -1,68 +1,0 @@
-package nextstep.security.filter;
-
-import jakarta.servlet.Filter;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.ServletRequest;
-import jakarta.servlet.ServletResponse;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.security.Authentication;
-import nextstep.security.AuthenticationManager;
-import nextstep.security.UsernamePasswordAuthenticationToken;
-import nextstep.security.context.SecurityContextHolder;
-import nextstep.security.exception.AuthenticationException;
-import nextstep.security.util.matcher.MvcRequestMatcher;
-import org.springframework.http.HttpMethod;
-
-import java.io.IOException;
-
-public class FormLoginAuthenticationFilter implements Filter {
-
-    public static final String SPRING_SECURITY_CONTEXT_KEY = "SPRING_SECURITY_CONTEXT";
-    private static final MvcRequestMatcher DEFAULT_REQUEST_MATCHER = new MvcRequestMatcher(HttpMethod.POST, "/login");
-
-    private final AuthenticationManager authenticationManager;
-
-    public FormLoginAuthenticationFilter(AuthenticationManager authenticationManager) {
-        this.authenticationManager = authenticationManager;
-    }
-
-    @Override
-    public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
-        if (!DEFAULT_REQUEST_MATCHER.matches((HttpServletRequest) servletRequest)) {
-            filterChain.doFilter(servletRequest, servletResponse);
-            return;
-        }
-
-        HttpServletRequest request = (HttpServletRequest) servletRequest;
-        HttpServletResponse response = (HttpServletResponse) servletResponse;
-
-        try {
-
-            Authentication authenticationResult = attemptAuthentication(request, response);
-            if (authenticationResult == null) {
-                return;
-            }
-
-            SecurityContextHolder.getContext().setAuthentication(authenticationResult);
-
-        } catch (Exception e) {
-            SecurityContextHolder.clearContext();
-            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-
-            return;
-        }
-
-        filterChain.doFilter(servletRequest, servletResponse);
-    }
-
-    private Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws AuthenticationException {
-
-        String username = request.getParameter("username");
-        String password = request.getParameter("password");
-        UsernamePasswordAuthenticationToken authRequest = UsernamePasswordAuthenticationToken.unauthenticated(username, password);
-
-        return authenticationManager.authenticate(authRequest);
-    }
-}
