--- src/main/java/nextstep/security/password/PasswordEncoder.java
@@ -1,8 +1,0 @@
-package nextstep.security.password;
-
-public interface PasswordEncoder {
-
-    String encode(String rawPassword);
-
-    boolean matches(String rawPassword, String encodedPassword);
-}
