--- src/main/java/nextstep/security/password/RawPasswordEncoder.java
@@ -1,14 +1,0 @@
-package nextstep.security.password;
-
-public class RawPasswordEncoder implements PasswordEncoder {
-
-    @Override
-    public String encode(String rawPassword) {
-        return rawPassword;
-    }
-
-    @Override
-    public boolean matches(String rawPassword, String encodedPassword) {
-        return rawPassword.equals(encodedPassword);
-    }
-}
