--- src/main/java/nextstep/security/util/matcher/AnyRequestMatcher.java
@@ -1,13 +1,0 @@
-package nextstep.security.util.matcher;
-
-import jakarta.servlet.http.HttpServletRequest;
-
-public class AnyRequestMatcher implements RequestMatcher {
-
-    public static final AnyRequestMatcher INSTANCE = new AnyRequestMatcher();
-
-    @Override
-    public boolean matches(HttpServletRequest request) {
-        return true;
-    }
-}
