--- src/main/java/nextstep/security/util/matcher/RequestMatcher.java
@@ -1,8 +1,0 @@
-package nextstep.security.util.matcher;
-
-import jakarta.servlet.http.HttpServletRequest;
-
-public interface RequestMatcher {
-
-    boolean matches(HttpServletRequest request);
-}
